# PERF: Series slicing