--- a/pandas/core/series.py
+++ b/pandas/core/series.py
@@ -981,6 +981,13 @@ class Series(base.IndexOpsMixin, NDFrame):  # type: ignore[misc]
                     # in the first level of our MultiIndex
                     return self._get_values_tuple(key)
 
+        if isinstance(key, slice):
+            # Do slice check before somewhat-costly is_bool_indexer
+            # _convert_slice_indexer to determine if this slice is positional
+            #  or label based, and if the latter, convert to positional
+            slobj = self.index._convert_slice_indexer(key, kind="getitem")
+            return self._slice(slobj)
+
         if is_iterator(key):
             key = list(key)
 
@@ -993,12 +1000,7 @@ class Series(base.IndexOpsMixin, NDFrame):  # type: ignore[misc]
 
     def _get_with(self, key):
         # other: fancy integer or otherwise
-        if isinstance(key, slice):
-            # _convert_slice_indexer to determine if this slice is positional
-            #  or label based, and if the latter, convert to positional
-            slobj = self.index._convert_slice_indexer(key, kind="getitem")
-            return self._slice(slobj)
-        elif isinstance(key, ABCDataFrame):
+        if isinstance(key, ABCDataFrame):
             raise TypeError(
                 "Indexing a Series with DataFrame is not "
                 "supported, use the appropriate DataFrame column"
@@ -1053,7 +1055,7 @@ class Series(base.IndexOpsMixin, NDFrame):  # type: ignore[misc]
 
     def _get_values(self, indexer: slice | npt.NDArray[np.bool_]) -> Series:
         new_mgr = self._mgr.getitem_mgr(indexer)
-        return self._constructor(new_mgr).__finalize__(self)
+        return self._constructor(new_mgr, fastpath=True).__finalize__(self)
 
     def _get_value(self, label, takeable: bool = False):
         """