```
import pandas as pd
ser = pd.Series(range(300_000))

%timeit ser[:30]
29.2 µs ± 749 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)  # <- main
15.7 µs ± 459 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)  # <- PR
```

BTW in this branch, `__finalize__` takes 13.6% of the runtime in that slicing.